fix(playerbot): Fix constructor calls and struct duplicates
Compilation errors resolved:
1. BotAI_HybridAI.cpp: Reverted include back to "HybridAIController.h"
   since the file is in the same directory (AI/)
2. CombatSpecializationTemplates.h: Added nullptr second parameter to
   MovementIntegration constructor call (requires PositionManager*)
3. IEquipmentManager.h: Moved EquipmentMetrics struct to namespace scope
   to avoid conflict with nested definition in EquipmentManager.h
4. EquipmentManager.h: Removed duplicate EquipmentMetrics nested struct,
   now uses namespace-scope version from IEquipmentManager.h
